Update queries & fragments, colocate related code close to each others
