+++ app/graphql/fragments.ts
@@ -1,41 +1,40 @@
-export const MEDIA_FRAGMENT = `#graphql
-  fragment Media on Media {
-    __typename
-    mediaContentType
-    alt
-    previewImage {
-      url
-    }
-    ... on MediaImage {
-      id
-      image {
+export const PRODUCT_OPTION_FRAGMENT = `#graphql
+  fragment ProductOption on ProductOption {
+    name
+    optionValues {
+      name
+      firstSelectableVariant {
         id
-        url
-        width
-        height
+        title
+        availableForSale
+        image {
+          url
+          altText
+          width
+          height
+        }
+        price {
+          amount
+          currencyCode
+        }
+        compareAtPrice {
+          amount
+          currencyCode
+        }
       }
-    }
-    ... on Video {
-      id
-      sources {
-        mimeType
-        url
+      swatch {
+        color
+        image {
+          mediaContentType
+          previewImage {
+            url
+            altText
+          }
+        }
       }
-    }
-    ... on Model3d {
-      id
-      sources {
-        mimeType
-        url
-      }
-    }
-    ... on ExternalVideo {
-      id
-      embedUrl
-      host
     }
   }
-`;
+` as const;
 
 export const PRODUCT_CARD_FRAGMENT = `#graphql
   fragment ProductCard on Product {
@@ -52,6 +51,9 @@ export const PRODUCT_CARD_FRAGMENT = `#graphql
         width
         height
       }
+    }
+    options {
+      ...ProductOption
     }
     badges: metafields(identifiers: [
       { namespace: "custom", key: "best_seller" }
@@ -100,53 +102,11 @@ export const PRODUCT_CARD_FRAGMENT = `#graphql
       }
     }
   }
+  ${PRODUCT_OPTION_FRAGMENT}
 `;
-
-export const FEATURED_COLLECTION_FRAGMENT = `#graphql
-  fragment FeaturedCollectionDetails on Collection {
-    id
-    title
-    handle
-    image {
-      altText
-      width
-      height
-      url
-    }
-  }
-`;
-
-export let COLLECTION_CONTENT_FRAGMENT = `#graphql
-  fragment CollectionContent on Collection {
-    id
-    handle
-    title
-    descriptionHtml
-    heading: metafield(namespace: "hero", key: "title") {
-      value
-    }
-    byline: metafield(namespace: "hero", key: "byline") {
-      value
-    }
-    cta: metafield(namespace: "hero", key: "cta") {
-      value
-    }
-    spread: metafield(namespace: "hero", key: "spread") {
-      reference {
-        ...Media
-      }
-    }
-    spreadSecondary: metafield(namespace: "hero", key: "spread_secondary") {
-      reference {
-        ...Media
-      }
-    }
-  }
-${MEDIA_FRAGMENT}
-` as const;
 
 export let PRODUCT_VARIANT_FRAGMENT = `#graphql
-  fragment ProductVariantFragment on ProductVariant {
+  fragment ProductVariant on ProductVariant {
     id
     availableForSale
     quantityAvailable
@@ -181,114 +141,3 @@ export let PRODUCT_VARIANT_FRAGMENT = `#graphql
     }
   }
 `;
-
-// NOTE: https://shopify.dev/docs/api/storefront/latest/queries/cart
-export const CART_QUERY_FRAGMENT = `#graphql
-  fragment Money on MoneyV2 {
-    currencyCode
-    amount
-  }
-  fragment CartLine on CartLine {
-    id
-    quantity
-    attributes {
-      key
-      value
-    }
-    cost {
-      totalAmount {
-        ...Money
-      }
-      amountPerQuantity {
-        ...Money
-      }
-      compareAtAmountPerQuantity {
-        ...Money
-      }
-    }
-    
-    sellingPlanAllocation {
-      sellingPlan {
-        name
-      }
-    }
-    merchandise {
-      ... on ProductVariant {
-        id
-        availableForSale
-        compareAtPrice {
-          ...Money
-        }
-        price {
-          ...Money
-        }
-        requiresShipping
-        title
-        image {
-          id
-          url
-          altText
-          width
-          height
-
-        }
-        product {
-          handle
-          title
-          id
-          vendor
-        }
-        selectedOptions {
-          name
-          value
-        }
-      }
-    }
-  }
-  fragment CartApiQuery on Cart {
-    updatedAt
-    id
-    checkoutUrl
-    totalQuantity
-    buyerIdentity {
-      countryCode
-      customer {
-        id
-        email
-        firstName
-        lastName
-        displayName
-      }
-      email
-      phone
-    }
-    lines(first: $numCartLines) {
-      nodes {
-        ...CartLine
-      }
-    }
-    cost {
-      subtotalAmount {
-        ...Money
-      }
-      totalAmount {
-        ...Money
-      }
-      totalDutyAmount {
-        ...Money
-      }
-      totalTaxAmount {
-        ...Money
-      }
-    }
-    note
-    attributes {
-      key
-      value
-    }
-    discountCodes {
-      code
-      applicable
-    }
-  }
-` as const;
