+++ app/graphql/queries.ts
@@ -1,158 +1,10 @@
 import {
-  COLLECTION_CONTENT_FRAGMENT,
-  MEDIA_FRAGMENT,
-  PRODUCT_CARD_FRAGMENT,
+  PRODUCT_OPTION_FRAGMENT,
   PRODUCT_VARIANT_FRAGMENT,
 } from "~/graphql/fragments";
 
-export const LAYOUT_QUERY = `#graphql
-  query layout(
-    $language: LanguageCode
-    $headerMenuHandle: String!
-    $footerMenuHandle: String!
-  ) @inContext(language: $language) {
-    shop {
-      ...Shop
-    }
-    headerMenu: menu(handle: $headerMenuHandle) {
-      ...Menu
-    }
-    footerMenu: menu(handle: $footerMenuHandle) {
-      ...Menu
-    }
-  }
-  fragment Shop on Shop {
-    id
-    name
-    description
-    primaryDomain {
-      url
-    }
-    brand {
-      logo {
-        image {
-          url
-        }
-      }
-    }
-  }
-  fragment MenuItem on MenuItem {
-    id
-    resourceId
-    resource {
-      ... on Collection {
-        image {
-          altText
-          height
-          id
-          url
-          width
-        }
-      }
-      ... on Product {
-        image: featuredImage {
-          altText
-          height
-          id
-          url
-          width
-        }
-      }
-    }
-    tags
-    title
-    type
-    url
-  }
-
-  fragment ChildMenuItem on MenuItem {
-    ...MenuItem
-  }
-  fragment ParentMenuItem2 on MenuItem {
-    ...MenuItem
-    items {
-      ...ChildMenuItem
-    }
-  }
-  fragment ParentMenuItem on MenuItem {
-    ...MenuItem
-    items {
-      ...ParentMenuItem2
-    }
-  }
-  fragment Menu on Menu {
-    id
-    items {
-      ...ParentMenuItem
-    }
-  }
-` as const;
-
-export const SWATCHES_CONFIGS_QUERY = `#graphql
-  query swatchesConfigs($type: String!) {
-    metaobjects(first: 100, type: $type) {
-      nodes {
-        id
-        fields {
-          key
-          value
-          reference {
-            ... on MediaImage {
-              id
-              image {
-                altText
-                id
-                url
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-` as const;
-
-export const SHOP_QUERY = `#graphql
-  query shopQuery($country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    shop {
-      name
-      description
-    }
-  }
-`;
-
-export const HOMEPAGE_SEO_QUERY = `#graphql
-  query seoCollectionContent($handle: String, $country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    hero: collection(handle: $handle) {
-      ...CollectionContent
-    }
-    shop {
-      name
-      description
-    }
-  }
-  ${COLLECTION_CONTENT_FRAGMENT}
-` as const;
-
-export const PRODUCT_INFO_QUERY = `#graphql
-  query ProductInfo(
-    $country: CountryCode
-    $language: LanguageCode
-    $handle: String!
-  ) @inContext(country: $country, language: $language) {
-    product(handle: $handle) {
-      id
-      title
-      vendor
-      handle
-    }
-  }
-` as const;
-
 export const PRODUCT_QUERY = `#graphql
-  query Product(
+  query product(
     $country: CountryCode
     $language: LanguageCode
     $handle: String!
@@ -168,34 +20,14 @@ export const PRODUCT_QUERY = `#graphql
       description
       summary: description(truncateAt: 200)
       options {
-        name
-        optionValues {
-          name
-          swatch {
-            color
-            image {
-              mediaContentType
-              previewImage {
-                url: url(transform: {maxWidth: 300})
-                altText
-                width
-                height
-              }
-            }
-          }
-        }
+        ...ProductOption
       }
       selectedVariant: variantBySelectedOptions(selectedOptions: $selectedOptions, ignoreUnknownOptions: true, caseInsensitiveMatch: true) {
-        ...ProductVariantFragment
+        ...ProductVariant
       }
       media(first: 7) {
         nodes {
           ...Media
-        }
-      }
-      variants(first: 1) {
-        nodes {
-          ...ProductVariantFragment
         }
       }
       seo {
@@ -218,369 +50,44 @@ export const PRODUCT_QUERY = `#graphql
       }
     }
   }
-  ${MEDIA_FRAGMENT}
+  fragment Media on Media {
+    __typename
+    mediaContentType
+    alt
+    previewImage {
+      url
+    }
+    ... on MediaImage {
+      id
+      image {
+        id
+        url
+        width
+        height
+      }
+    }
+    ... on Video {
+      id
+      sources {
+        mimeType
+        url
+      }
+    }
+    ... on Model3d {
+      id
+      sources {
+        mimeType
+        url
+      }
+    }
+    ... on ExternalVideo {
+      id
+      embedUrl
+      host
+    }
+  }
   ${PRODUCT_VARIANT_FRAGMENT}
-` as const;
-
-export const RECOMMENDED_PRODUCTS_QUERY = `#graphql
-  query productRecommendations(
-    $productId: ID!
-    $count: Int
-    $country: CountryCode
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    recommended: productRecommendations(productId: $productId) {
-      ...ProductCard
-    }
-    additional: products(first: $count, sortKey: BEST_SELLING) {
-      nodes {
-        ...ProductCard
-      }
-    }
-  }
-  ${PRODUCT_CARD_FRAGMENT}
-` as const;
-
-export const COLLECTION_INFO_QUERY = `#graphql
-  query CollectionInfo(
-    $handle: String!,
-    $country: CountryCode,
-    $language: LanguageCode
-  ) @inContext(country: $country, language: $language) {
-    collection(handle: $handle) {
-      id
-      handle
-      title
-      description
-      seo {
-        description
-        title
-      }
-      image {
-        id
-        url
-        width
-        height
-        altText
-      }
-    }
-  }
-` as const;
-
-export const COLLECTION_QUERY = `#graphql
-  query CollectionDetails(
-    $handle: String!
-    $country: CountryCode
-    $language: LanguageCode
-    $filters: [ProductFilter!]
-    $sortKey: ProductCollectionSortKeys!
-    $reverse: Boolean
-    $first: Int
-    $last: Int
-    $startCursor: String
-    $endCursor: String
-    $customBannerNamespace: String!
-    $customBannerKey: String!
-  ) @inContext(country: $country, language: $language) {
-    collection(handle: $handle) {
-      id
-      handle
-      title
-      description
-      seo {
-        description
-        title
-      }
-      metafield(namespace: $customBannerNamespace, key: $customBannerKey) {
-        id
-        type
-        description
-        value
-        reference {
-          ... on MediaImage {
-            image {
-              id
-              url
-            }
-          }
-        }
-      }
-      image {
-        id
-        url
-        width
-        height
-        altText
-      }
-      products(
-        first: $first,
-        last: $last,
-        before: $startCursor,
-        after: $endCursor,
-        filters: $filters,
-        sortKey: $sortKey,
-        reverse: $reverse
-      ) {
-        filters {
-          id
-          label
-          type
-          values {
-            id
-            label
-            count
-            input
-          }
-        }
-        nodes {
-          ...ProductCard
-        }
-        pageInfo {
-          hasPreviousPage
-          hasNextPage
-          endCursor
-          startCursor
-        }
-      }
-      highestPriceProduct: products(first: 1, sortKey: PRICE, reverse: true) {
-        nodes {
-          id
-          title
-          handle
-          priceRange {
-            minVariantPrice {
-              amount
-              currencyCode
-            }
-            maxVariantPrice {
-              amount
-              currencyCode
-            }
-          }
-        }
-      }
-      lowestPriceProduct: products(first: 1, sortKey: PRICE) {
-        nodes {
-          id
-          title
-          handle
-          priceRange {
-            minVariantPrice {
-              amount
-              currencyCode
-            }
-            maxVariantPrice {
-              amount
-              currencyCode
-            }
-          }
-        }
-      }
-    }
-    collections(first: 100) {
-      edges {
-        node {
-          title
-          handle
-        }
-      }
-    }
-  }
-  ${PRODUCT_CARD_FRAGMENT}
-` as const;
-
-export const COLLECTIONS_QUERY = `#graphql
-  query Collections(
-    $country: CountryCode
-    $language: LanguageCode
-    $first: Int
-    $last: Int
-    $startCursor: String
-    $endCursor: String
-  ) @inContext(country: $country, language: $language) {
-    collections(first: $first, last: $last, before: $startCursor, after: $endCursor) {
-      nodes {
-        id
-        title
-        description
-        handle
-        seo {
-          description
-          title
-        }
-        image {
-          id
-          url
-          width
-          height
-          altText
-        }
-      }
-      pageInfo {
-        hasPreviousPage
-        hasNextPage
-        startCursor
-        endCursor
-      }
-    }
-  }
-`;
-
-export const SEARCH_QUERY = `#graphql
-  query PaginatedProductsSearch(
-    $country: CountryCode
-    $endCursor: String
-    $first: Int
-    $language: LanguageCode
-    $last: Int
-    $searchTerm: String
-    $startCursor: String
-  ) @inContext(country: $country, language: $language) {
-    products(
-      first: $first,
-      last: $last,
-      before: $startCursor,
-      after: $endCursor,
-      sortKey: RELEVANCE,
-      query: $searchTerm
-    ) {
-      nodes {
-        ...ProductCard
-      }
-      pageInfo {
-        startCursor
-        endCursor
-        hasNextPage
-        hasPreviousPage
-      }
-    }
-  }
-
-  ${PRODUCT_CARD_FRAGMENT}
-` as const;
-
-export const BLOGS_QUERY = `#graphql
-  query Blog(
-    $language: LanguageCode
-    $blogHandle: String!
-    $pageBy: Int!
-    $cursor: String
-  ) @inContext(language: $language) {
-    blog(handle: $blogHandle) {
-      title
-      handle
-      seo {
-        title
-        description
-      }
-      articles(first: $pageBy, after: $cursor) {
-        edges {
-          node {
-            ...Article
-          }
-        }
-      }
-    }
-  }
-
-  fragment Article on Article {
-    author: authorV2 {
-      name
-    }
-    contentHtml
-    excerpt
-    excerptHtml
-    handle
-    id
-    image {
-      id
-      altText
-      url
-      width
-      height
-    }
-    publishedAt
-    title
-  }
-`;
-
-export const ARTICLE_QUERY = `#graphql
-  query ArticleDetails(
-    $language: LanguageCode
-    $blogHandle: String!
-    $articleHandle: String!
-  ) @inContext(language: $language) {
-    blog(handle: $blogHandle) {
-      articleByHandle(handle: $articleHandle) {
-        title
-        contentHtml
-        publishedAt
-        tags
-        author: authorV2 {
-          name
-        }
-        image {
-          id
-          altText
-          url
-          width
-          height
-        }
-        seo {
-          description
-          title
-        }
-      }
-      articles (first: 20) {
-        nodes {
-            ...Article
-        }
-      }
-    }
-  }
-  fragment Article on Article {
-    author: authorV2 {
-      name
-    }
-    contentHtml
-    excerpt
-    excerptHtml
-    handle
-    id
-    image {
-      id
-      altText
-      url
-      width
-      height
-    }
-    publishedAt
-    title
-  }
-`;
-
-export const ALL_PRODUCTS_QUERY = `#graphql
-  query AllProducts(
-    $country: CountryCode
-    $language: LanguageCode
-    $first: Int
-    $last: Int
-    $startCursor: String
-    $endCursor: String
-  ) @inContext(country: $country, language: $language) {
-    products(first: $first, last: $last, before: $startCursor, after: $endCursor) {
-      nodes {
-        ...ProductCard
-      }
-      pageInfo {
-        hasPreviousPage
-        hasNextPage
-        startCursor
-        endCursor
-      }
-    }
-  }
-  ${PRODUCT_CARD_FRAGMENT}
+  ${PRODUCT_OPTION_FRAGMENT}
 ` as const;
 
 export const VARIANTS_QUERY = `#graphql
@@ -592,7 +99,7 @@ export const VARIANTS_QUERY = `#graphql
     product(handle: $handle) {
       variants(first: 250) {
         nodes {
-          ...ProductVariantFragment
+          ...ProductVariant
         }
       }
     }
