+++ app/routes/($locale).blogs.$blogHandle.$articleHandle.tsx
@@ -6,7 +6,6 @@ import type { RouteLoaderArgs, WeaverseClient } from "@weaverse/hydrogen";
 import invariant from "tiny-invariant";
 import type { ArticleDetailsQuery } from "storefront-api.generated";
 import { routeHeaders } from "~/utils/cache";
-import { ARTICLE_QUERY } from "~/graphql/queries";
 import { seoPayload } from "~/utils/seo.server";
 import { WeaverseContent } from "~/weaverse";
 
@@ -67,3 +66,58 @@ export let meta: MetaFunction<typeof loader> = ({ data }) => {
 export default function Article() {
   return <WeaverseContent />;
 }
+
+const ARTICLE_QUERY = `#graphql
+  query article(
+    $language: LanguageCode
+    $blogHandle: String!
+    $articleHandle: String!
+  ) @inContext(language: $language) {
+    blog(handle: $blogHandle) {
+      articleByHandle(handle: $articleHandle) {
+        title
+        contentHtml
+        publishedAt
+        tags
+        author: authorV2 {
+          name
+        }
+        image {
+          id
+          altText
+          url
+          width
+          height
+        }
+        seo {
+          description
+          title
+        }
+      }
+      articles (first: 20) {
+        nodes {
+            ...Article
+        }
+      }
+    }
+  }
+  fragment Article on Article {
+    author: authorV2 {
+      name
+    }
+    contentHtml
+    excerpt
+    excerptHtml
+    handle
+    id
+    image {
+      id
+      altText
+      url
+      width
+      height
+    }
+    publishedAt
+    title
+  }
+` as const;
